requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
google-genai
google-api-core
